+++ CHANGES.md
@@ -2,8 +2,13 @@
 
 ## [Unreleased]
 
+### Removed
+
+* Remove deprecated `context` property in `types.stac.ItemCollection` model
+
 ### Added
 
+* Add `numberMatched` and `numberReturned` properties in `types.stac.ItemCollection` model
 * Add `numberMatched` and `numberReturned` properties in `types.stac.Collections` model
 
 ## [3.0.3] - 2024-10-09