+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [3.0.4] - 2025-01-08
+
 ### Removed
 
 * Remove deprecated `context` property in `types.stac.ItemCollection` model
@@ -507,7 +509,8 @@ Full changelog: https://stac-utils.github.io/stac-fastapi/migrations/v3.0.0/#cha
 
 * First PyPi release!
 
-[Unreleased]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.3..main>
+[Unreleased]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.4..main>
+[3.0.4]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.3..3.0.4>
 [3.0.3]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.2..3.0.3>
 [3.0.2]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.1..3.0.2>
 [3.0.1]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.0..3.0.1>